Marketplace: Move components into it's own modules (#66060)
* Refactor plugins to components
